Created selectionSort.c
